+++ userspace/lib/include/fcntl.h
@@ -12,6 +12,10 @@
 #define O_NONBLOCK 0x0800
 #define O_DIRECTORY 0x10000
 
+#define FD_CLOEXEC 1
+
+#define F_GETFD 1
+#define F_SETFD 2
 #define F_GETFL 3
 #define F_SETFL 4